+++ components/remote-controls/grid/saved/impl/src/main/kotlin/com/flipperdevices/remotecontrols/impl/grid/local/composable/LocalGridComposable.kt
@@ -14,10 +14,12 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
-import com.flipperdevices.bridge.synchronization.api.SynchronizationState
 import com.flipperdevices.core.ui.theme.LocalPalletV2
 import com.flipperdevices.ifrmvp.core.ui.layout.shared.SharedTopBar
+import com.flipperdevices.infrared.api.InfraredConnectionApi.InfraredEmulateState
+import com.flipperdevices.remotecontrols.grid.saved.impl.R
 import com.flipperdevices.remotecontrols.impl.grid.local.api.LocalGridScreenDecomposeComponent
 import com.flipperdevices.remotecontrols.impl.grid.local.composable.components.ComposableInfraredDropDown
 import com.flipperdevices.remotecontrols.impl.grid.local.composable.components.ComposableSynchronizationNotification
@@ -46,7 +48,11 @@ fun LocalGridComposable(
         modifier = modifier,
         topBar = {
             (model as? LocalGridComponent.Model.Loaded)?.let { loadedModel ->
-                SharedTopBar(onBackClick = localGridComponent::pop) {
+                SharedTopBar(
+                    onBackClick = localGridComponent::pop,
+                    title = loadedModel.keyPath.path.nameWithoutExtension,
+                    subtitle = stringResource(R.string.remote_subtitle)
+                ) {
                     ComposableInfraredDropDown(
                         onRename = {
                             localGridComponent.onRename {
@@ -62,11 +68,14 @@ fun LocalGridComposable(
                         },
                         onRemoteInfo = {
                             onCallback.invoke(
-                                LocalGridScreenDecomposeComponent.Callback.ViewRemoteInfo(loadedModel.keyPath)
+                                LocalGridScreenDecomposeComponent.Callback.ViewRemoteInfo(
+                                    loadedModel.keyPath
+                                )
                             )
                         },
                         onShare = onShare,
-                        emulatingInProgress = loadedModel.emulatedKey != null
+                        isEmulating = loadedModel.emulatedKey != null,
+                        isConnected = loadedModel.isConnected
                     )
                 }
             }
@@ -87,8 +96,8 @@ fun LocalGridComposable(
                 contentAlignment = Alignment.BottomCenter
             ) {
                 val state = (model as? LocalGridComponent.Model.Loaded)
-                    ?.synchronizationState
-                    ?: SynchronizationState.NotStarted
+                    ?.connectionState
+                    ?: InfraredEmulateState.ALL_GOOD
                 ComposableSynchronizationNotification(state)
             }
         }