+++ components/rootscreen/api/src/main/kotlin/com/flipperdevices/rootscreen/model/RootScreenConfig.kt
@@ -39,5 +39,8 @@ sealed class RootScreenConfig {
     data object RemoteControls : RootScreenConfig()
 
     @Serializable
-    data class ServerRemoteControl(val infraredFileId: Long) : RootScreenConfig()
+    data class ServerRemoteControl(
+        val infraredFileId: Long,
+        val remoteName: String
+    ) : RootScreenConfig()
 }